Fix broken tests after moving BioentityPropertyName to the new package
